backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.35
python-multipart>=0.0.12
pillow>=10.4.0
pytesseract>=0.3.13
python-telegram-bot>=21.6
joblib>=1.4.2
scikit-learn>=1.3.0
